refactor(exec): extract runDetached helper for reuse
